packages/shipped-it/pyproject.toml
[project]
name = "shipped-it"
version = "0.1.0"
description = "Core of shipped-it"
readme = "README.md"
authors = [
    { name = "Kazuya Takei", email = "myself@attakei.net" }
]
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tweepy>=4.16.0",
]

[project.scripts]
shipped = "shipped_it.cli:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
